test listviewEx

+++ test/test_net/test_net/test_net/cmd_Form.cs
@@ -13,9 +13,50 @@ namespace test_net
     public delegate void add_send_command(string str);
     public partial class cmd_Form : Form
     {
+        struct LST_SIZE
+        {
+            public int INFO_WIDTH;
+            public int FORMAT_WIDTH;
+            public int DATA_WIDTH;
+            public int DLY_WIDTH;
+            public int BTN_WIDTH;
+            public int LIST_HEIGHT;
+        }
+        struct LST_NAME
+        {
+            public string INFO_NAME;
+            public string FORMAT_NAME;
+            public string DATA_NAME;
+            public string DLY_NAME;
+            public string BTN_NAME;
+        }
+
+        int item_num = 0;
+        LST_SIZE lst_size;
+        LST_NAME lst_name;
         public cmd_Form()
         {
             InitializeComponent();
+            cmd_load();
+        }
+
+        private void cmd_load()
+        {
+            //添加控件大小信息
+            lst_size.INFO_WIDTH = columnHeader2.Width;
+            lst_size.FORMAT_WIDTH = columnHeader3.Width;
+            lst_size.DATA_WIDTH = columnHeader4.Width;
+            lst_size.DLY_WIDTH = columnHeader5.Width;
+            lst_size.BTN_WIDTH = columnHeader6.Width;
+            lst_size.LIST_HEIGHT = 26;
+            //添加控件名称信息
+            lst_name.INFO_NAME = "Info:";
+            lst_name.FORMAT_NAME = "Fmt:";
+            lst_name.DATA_NAME = "Data:";
+            lst_name.DLY_NAME = "Dly:";
+            lst_name.BTN_NAME = "Btn:";
+
+            panel1.VerticalScroll.Value = panel1.VerticalScroll.Maximum;
         }
         public event add_send_command send_command;
         private void cmd_send_button_Click(object sender, EventArgs e)
@@ -24,7 +65,6 @@ namespace test_net
             //this.Close();
         }
 
-        int item_num = 0;
         private void add_item_button_Click(object sender, EventArgs e)
         {
             add_item(item_num);
@@ -33,34 +73,46 @@ namespace test_net
 
         private void add_item(int i)
         {
+            //先将焦点移到最上方
+            //panel1.AutoScrollPosition = new Point(0, 0);
+            //command_listView.AutoScrollOffset = new Point(0, 0);
+
             listview(true);
-            info_text(i);
-            format_choice(i);
-            send_text(i);
-            sendDelay_text(i);
-            send_button(i);
+            
+
+            //添加控件
+            //add_info_text(i, lst_size);
+            //add_format_choice(i, lst_size);
+            //add_send_text(i, lst_size);
+            //add_sendDelay_text(i, lst_size);
+            add_send_button(i, lst_size);
+            //将滚动条移至最后
+            //Point newPoint = new Point(0, this.panel1.Height - panel1.AutoScrollPosition.Y);
+            //panel1.AutoScrollPosition = newPoint;
+
+            //this.command_listView.Items[this.command_listView.Items.Count - 1].EnsureVisible();
+            //command_listView.AutoScrollOffset = new Point(0, lst_size.LIST_HEIGHT*i);
         }
 
         private void delete_item(int i)
         {        
             //删除选中项
-            remove_item(i);
-            //将其余 item 上移
-
-
-            listview(false);     
+            remove_item(i);   
         }
+
         private void listview(bool act)
         {
-            ListViewItem[] cmdlst_item = new ListViewItem[1];
-
-            ImageList imgList = new ImageList();
-            imgList.ImageSize = new Size(1, 25);//分别是宽和高
-            command_listView.SmallImageList = imgList;
-
-            cmdlst_item[0] = new ListViewItem(new string[] { "", "", "", "", "" });
             if (act == true)
             {
+                ListViewItem[] cmdlst_item = new ListViewItem[1];
+
+                //设置行高
+                ImageList imgList = new ImageList();
+                imgList.ImageSize = new Size(1, 25);
+                command_listView.SmallImageList = imgList;
+
+                //添加元素
+                cmdlst_item[0] = new ListViewItem(new string[] { "", item_num.ToString(), "2", "3", "4", "5" });
                 this.command_listView.Items.AddRange(cmdlst_item);
             }
             else
@@ -69,64 +121,65 @@ namespace test_net
             }
         }
 
-        enum item_info{  TEXTBOX_INFO, COMBOBOX, TEXTBOX_DATA, TEXTBOX_DELAY, BUTTON_SEND };
+        enum item_info{ HEAD, TEXTBOX_INFO, COMBOBOX, TEXTBOX_DATA, TEXTBOX_DELAY, BUTTON_SEND };
         
-        private void send_button(int i)
+        private void add_send_button(int i, LST_SIZE lst)
         {
             Button m_button = new Button();
             //按钮的位置
             m_button.Click += new EventHandler(item_send_button_Click);
             m_button.Text = "发送"+ i.ToString();
-            m_button.Tag = "Btn:" + i.ToString();
+            m_button.Tag = lst_name.BTN_NAME + i.ToString();
             m_button.BackColor = Color.LightBlue;
             m_button.Font = new Font("黑体", 7);
-            m_button.Size = new Size(this.command_listView.Items[i].SubItems[(int)item_info.BUTTON_SEND].Bounds.Width - 1, this.command_listView.Items[i].SubItems[(int)item_info.BUTTON_SEND].Bounds.Height - 1);
-            m_button.Location = new Point(this.command_listView.Items[i].SubItems[(int)item_info.BUTTON_SEND].Bounds.Left - 1, this.command_listView.Items[i].SubItems[(int)item_info.BUTTON_SEND].Bounds.Top - 1);
-            this.command_listView.Controls.Add(m_button);
+            m_button.Size = new Size(lst.BTN_WIDTH, lst.LIST_HEIGHT);
+            m_button.Location = new Point(1+lst.INFO_WIDTH+ lst.FORMAT_WIDTH+ lst.DATA_WIDTH+ lst.DLY_WIDTH, lst.LIST_HEIGHT * i);
+            //this.command_listView.Controls.Add(m_button);
+            command_listView.AddEmbeddedControl(m_button, 3, 0);
         }
 
-        private void sendDelay_text(int i)
+        private void add_sendDelay_text(int i, LST_SIZE lst)
         {
             TextBox delay_parm = new TextBox();
             //延时框的位置
-            delay_parm.Tag = "Dly:" + i.ToString();
+            delay_parm.Tag = lst_name.DLY_NAME + i.ToString();
             delay_parm.Text = "0";
-            delay_parm.Size = new Size(this.command_listView.Items[i].SubItems[3].Bounds.Width - 1, this.command_listView.Items[i].SubItems[3].Bounds.Height - 1);
-            delay_parm.Location = new Point(this.command_listView.Items[i].SubItems[3].Bounds.Left - 1, this.command_listView.Items[i].SubItems[3].Bounds.Top - 1);
+            delay_parm.Size = new Size(lst.DLY_WIDTH, lst.LIST_HEIGHT);
+            delay_parm.Location = new Point(1+lst.INFO_WIDTH + lst.FORMAT_WIDTH + lst.DATA_WIDTH , lst.LIST_HEIGHT * i);
             this.command_listView.Controls.Add(delay_parm);
         }
 
-        private void send_text(int i)
+        private void add_send_text(int i, LST_SIZE lst)
         {
             TextBox send_data = new TextBox();
             //发送框的位置
-            send_data.Tag = "Data:" + i.ToString();
-            send_data.Size = new Size(this.command_listView.Items[i].SubItems[2].Bounds.Width - 1, this.command_listView.Items[i].SubItems[2].Bounds.Height - 1);
-            send_data.Location = new Point(this.command_listView.Items[i].SubItems[2].Bounds.Left - 1, this.command_listView.Items[i].SubItems[2].Bounds.Top - 1);
+            send_data.Tag = lst_name.DATA_NAME + i.ToString();
+            send_data.Text = item_num.ToString();
+            send_data.Size = new Size(lst.DATA_WIDTH, lst.LIST_HEIGHT);
+            send_data.Location = new Point(1 + lst.INFO_WIDTH + lst.FORMAT_WIDTH, lst.LIST_HEIGHT * i);
             this.command_listView.Controls.Add(send_data);
         }
 
-        private void format_choice(int i)
+        private void add_format_choice(int i, LST_SIZE lst)
         {
             ComboBox send_format = new ComboBox();
-            send_format.Tag = "Cmb:"+ i.ToString();
+            send_format.Tag = lst_name.FORMAT_NAME + i.ToString();
             send_format.Items.Add("十六进制");
             send_format.Items.Add("字符串");
             send_format.SelectedIndex = 0;
-            send_format.Size = new Size(this.command_listView.Items[i].SubItems[1].Bounds.Width - 1, this.command_listView.Items[i].SubItems[1].Bounds.Height - 1);
-            send_format.Location = new Point(this.command_listView.Items[i].SubItems[1].Bounds.Left - 1, this.command_listView.Items[i].SubItems[1].Bounds.Top - 1);
+            send_format.Size = new Size(lst.FORMAT_WIDTH, lst.LIST_HEIGHT);
+            send_format.Location = new Point(1 + lst.INFO_WIDTH, lst.LIST_HEIGHT * i);
             this.command_listView.Controls.Add(send_format);
         }
 
-        private void info_text(int i)
+        private void add_info_text(int i, LST_SIZE lst)
         {
             TextBox add_info = new TextBox();
-            add_info.Tag = "Info:" + i.ToString();
-            add_info.Size = new Size(this.command_listView.Items[i].SubItems[1].Bounds.Width - 35, this.command_listView.Items[i].SubItems[1].Bounds.Height -1);
-            add_info.Location = new Point(this.command_listView.Items[i].SubItems[0].Bounds.Left, this.command_listView.Items[i].SubItems[0].Bounds.Top - 1);
+            add_info.Tag = lst_name.INFO_NAME + i.ToString();
+            add_info.Size = new Size(lst.INFO_WIDTH, lst.LIST_HEIGHT);
+            add_info.Location = new Point(1, lst.LIST_HEIGHT * i);
             this.command_listView.Controls.Add(add_info);
         }
-
         private void item_send_button_Click(object sender, EventArgs e)
         {
             Button temp_btn = (Button)sender;
@@ -145,7 +198,7 @@ namespace test_net
         {
             Control err = null;
             int off_set = 0;
-            foreach (Control item in this.command_listView.Controls)
+            foreach (Control item in this.panel1.Controls)
             {
                 if (item_name == (int)item_info.BUTTON_SEND)
                 {
@@ -156,8 +209,8 @@ namespace test_net
                         if (off_set != 0)
                         {
                             string num = test.Tag.ToString().Substring(off_set + 1);
-                            string id = test.Tag.ToString().Substring(0, off_set);
-                            if (id.Equals("Btn"))
+                            string id = test.Tag.ToString().Substring(0, off_set+1);
+                            if (id.Equals(lst_name.BTN_NAME))
                             {
                                 if (num.Equals(item_num.ToString()))
                                 {
@@ -176,8 +229,8 @@ namespace test_net
                         if (off_set != 0)
                         {
                             string num = test.Tag.ToString().Substring(off_set + 1);
-                            string id = test.Tag.ToString().Substring(0, off_set);
-                            if (id.Equals("Cmb"))
+                            string id = test.Tag.ToString().Substring(0, off_set+1);
+                            if (id.Equals(lst_name.FORMAT_NAME))
                             {
                                 if (num.Equals(item_num.ToString()))
                                 {
@@ -196,8 +249,8 @@ namespace test_net
                         if (off_set != 0)
                         {
                             string num = test.Tag.ToString().Substring(off_set + 1);
-                            string id = test.Tag.ToString().Substring(0, off_set);
-                            if (id.Equals("Data"))
+                            string id = test.Tag.ToString().Substring(0, off_set+1);
+                            if (id.Equals(lst_name.DATA_NAME))
                             {
                                 if (num.Equals(item_num.ToString()))
                                 {
@@ -216,8 +269,8 @@ namespace test_net
                         if (off_set != 0)
                         {
                             string num = test.Tag.ToString().Substring(off_set + 1);
-                            string id = test.Tag.ToString().Substring(0, off_set);
-                            if (id.Equals("Dly"))
+                            string id = test.Tag.ToString().Substring(0, off_set+1);
+                            if (id.Equals(lst_name.DLY_NAME))
                             {
                                 if (num.Equals(item_num.ToString()))
                                 {
@@ -236,8 +289,8 @@ namespace test_net
                         if (off_set != 0)
                         {
                             string num = test.Tag.ToString().Substring(off_set + 1);
-                            string id = test.Tag.ToString().Substring(0, off_set);
-                            if (id.Equals("Info"))
+                            string id = test.Tag.ToString().Substring(0, off_set+1);
+                            if (id.Equals(lst_name.INFO_NAME))
                             {
                                 if (num.Equals(item_num.ToString()))
                                 {
@@ -256,7 +309,7 @@ namespace test_net
             for(int i = (int)item_info.TEXTBOX_INFO; i<= (int)item_info.BUTTON_SEND; i++)
             {
                 item = m_find_item(idx_num, i);
-                this.command_listView.Controls.Remove(item);
+                this.panel1.Controls.Remove(item);
             }
         }
 
@@ -270,36 +323,39 @@ namespace test_net
                     item = m_find_item(j + 1, i);
                     if (i == (int)item_info.TEXTBOX_INFO)
                     {
-                        item.Tag = "Info:" + j.ToString();
+                        item.Tag = lst_name.INFO_NAME + j.ToString();
+                        item.Location = new Point();
                     }
                     else if (i == (int)item_info.COMBOBOX)
                     {
-                        item.Tag = "Cmb:" + j.ToString();
+                        item.Tag = lst_name.FORMAT_NAME + j.ToString();
                     }
                     else if (i == (int)item_info.TEXTBOX_DATA)
                     {
-                        item.Tag = "Data:" + j.ToString();
+                        item.Tag = lst_name.DATA_NAME + j.ToString();
                     }
                     else if (i == (int)item_info.TEXTBOX_DELAY)
                     {
-                        item.Tag = "Dly:" + j.ToString();
+                        item.Tag = lst_name.DLY_NAME + j.ToString();
                     }
                     else if (i == (int)item_info.BUTTON_SEND)
                     {
-                        item.Tag = "Btn:" + j.ToString();
+                        item.Tag = lst_name.BTN_NAME + j.ToString();
                     }
-                    item.Size = new Size(this.command_listView.Items[j].SubItems[i].Bounds.Width - 1, this.command_listView.Items[j].SubItems[i].Bounds.Height - 1);
-                    item.Location = new Point(this.command_listView.Items[j].SubItems[i].Bounds.Left - 1, this.command_listView.Items[j].SubItems[i].Bounds.Top - 1);
                 }
             }
         }
 
         private void del_item_button_Click(object sender, EventArgs e)
         {
-            int idx_num = command_listView.SelectedItems[0].Index;
-            delete_item(idx_num);
-            move_up_item(idx_num, item_num);
-            item_num--;
+            int idx_num;
+            if (command_listView.SelectedItems[0] != null)
+            {
+                idx_num = command_listView.SelectedItems[0].Index;
+                delete_item(idx_num);
+                move_up_item(idx_num, item_num);
+                item_num--;
+            }
         }
     }
 }
